updateUser route made
