+++ online-shopping-website/backend/src/prismaFunctions/prismaFunctions.ts
@@ -1,0 +1,1 @@
+import prisma from "./PrismaClient";
